Add helpers to avoid code repetition for attributes. Remove attr files.
